frontend/requirements.txt
altair==5.2.0
asn1crypto==1.5.1
attrs==23.2.0
blinker==1.7.0
branca==0.7.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
extra-streamlit-components==0.1.60
folium==0.16.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pg8000==1.30.4
pillow==10.2.0
protobuf==4.25.2
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
scramp==1.4.4
six==1.16.0
smmap==5.0.1
streamlit==1.31.0
streamlit-folium==0.18.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
xyzservices==2023.10.1
zipp==3.17.0
plotly
matplotlib